api/Facial Recognition/requirements.txt
fastapi
uvicorn
face_recognition
python-multipart
pillow
